Creating types for record array proxy
